+++ packages/jira-adapter/src/change_validators/index.ts
@@ -25,7 +25,7 @@ import { JiraConfig } from '../config/config'
 import { projectDeletionValidator } from './project_deletion'
 import { statusValidator } from './status'
 import { privateApiValidator } from './private_api'
-import { workflowValidator } from './workflow'
+import { workflowValidator } from './workflows/workflow'
 import { dashboardGadgetsValidator } from './dashboard_gadgets'
 import { dashboardLayoutValidator } from './dashboard_layout'
 import { permissionTypeValidator } from './permission_type'
