+++ packages/jira-adapter/src/change_validators/workflows/workflow.ts
@@ -15,7 +15,7 @@
 */
 import { ChangeValidator, getChangeData, isInstanceChange, isRemovalOrModificationChange, SeverityLevel } from '@salto-io/adapter-api'
 import { collections } from '@salto-io/lowerdash'
-import { WORKFLOW_TYPE_NAME } from '../constants'
+import { WORKFLOW_TYPE_NAME } from '../../constants'
 
 const { awu } = collections.asynciterable
